Aplicacion/requirements.txt
Django==3.2.4
Markdown==3.3.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
